[nanda] fix bug authorize
