+++ server/middlewares/isAuthorizeTodo.js
@@ -1,0 +1,18 @@
+const { Todo } = require('../models');
+
+const isAuthorize = async(req, res, next) => {
+    try {
+        const todo = await Todo.findByPk(req.params.id);
+
+
+        if (!todo || todo.userId !== req.user.id) {
+            return next({ name: "unauthorize" })
+        }
+
+        return next();
+    } catch (err) {
+        return next(err)
+    }
+}
+
+module.exports = isAuthorize
